chore: inline createFragment into renderer-factory
